requirements.txt
aioredis==2.0.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
asyncio==3.4.3
bidict==0.22.1
certifi==2022.12.7
chardet==4.0.0
click==8.1.7
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.115.5
h11==0.14.0
httptools==0.6.4
idna==2.10
loguru==0.7.2
neo-api-client==1.2.0
numpy==1.24.2
pandas==2.0.0
pydantic==2.10.2
pydantic_core==2.27.1
pyjsparser==2.7.1
PyJWT==2.6.0
pymongo==4.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
requests==2.25.1
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.14
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.0
websocket-client==1.5.1
websockets==14.1
